Add files via upload
The assessment is on the topic "Performing Descriptive Statistics with Python" and this was executed using the Jupyter Notebook. Questions numbered 1 to 7  require a logical thinking and the answers were derived from the information given in the questions. It was merely based on the interpretation analysis and questions from 8 to 20 were based on the "Bank Personal Modelling Loan (1) " dataset wherein measures of central tendency, dispersion and certain plots were to be visualized for more insights.